Docs Audit - Utilities (#1071)
* Initial commit.

* Typos and grammar.

* typos: abitrary -> arbitrary

* typos and grammar

* Updated lightswitch setter import references

Credit to Sarenor for the catch :+1:

* Updates lightswitch getter import reference

* typos and grammar

* typo and improved consistency

* grammar fix

Please forgive me, senpai. I don't want to nitpick but that is what Doc reviews are by nature.

The singleton refers to the unique pattern, it is not one of many singletons ( ie. a singleton of the singletons ) so it is the singleton. Also, the wiki refers to it using "the" instead of "a". I can't believe I'm using wikipedia as an appeal to authority.

* Removed empty space

* typo and removed empty space

* Describe modal widths preset classes

---------

Co-authored-by: endigo9740 <gundamx9740@gmail.com>